fix/seperate-utils-files
While this doesn't allow for pixel perfect detection on windows it
atleast allows the project to build
